## Summary

Completes the `int` → `int64` migration for nodeID values that was
started in 66e8c8d but left 4 locations using `int`. Node IDs are
derived from `keccak256(operatorAddress)` and occupy 63 bits. Using
`int` (which is 32-bit on some platforms and in some map/serialization
contexts) silently truncates these values, causing map key collisions in
the share, commitment, and acknowledgement data structures used by DKG
and reshare protocols.

This is a correctness issue, not just a style concern: two distinct
operators whose keccak256-derived IDs differ only in the upper 32 bits
would map to the same key, causing one operator's share to overwrite
another's. In a threshold cryptography system, this breaks polynomial
secret sharing — the share verification equation `s * G2 == Σ(C_k *
j^k)` evaluates at the wrong point, and Lagrange interpolation during
key recovery uses the wrong coefficients.

## Changes

| File | What changed |
|---|---|
| `pkg/types/types.go` | `CompletionSignature.NodeID`: `int` → `int64` |
| `pkg/bls/polynomial.go` | `VerifyShare()`: `nodeID int` → `int64`;
`GenerateShares()`: `[]int` → `[]int64`, returns `map[int64]` |
| `pkg/crypto/bls.go` | `VerifyShareWithCommitments()`: `nodeID int` →
`int64` |
| `pkg/reshare/reshare.go` | `CreateCompletionSignature()`: `nodeID int`
→ `int64` |
| `pkg/bls/bls_test.go` | Updated test literals to `int64` |
| `pkg/bls/polynomial_fuzz_test.go` | Updated to `int64`; removed
now-unnecessary `toInt64ShareMap` helper |
| `pkg/reshare/reshare_test.go` | Updated test literal to `int64` |

+++ pkg/bls/bls_test.go
@@ -226,7 +226,7 @@ func testPolynomialSecretSharing(t *testing.T) {
 	}
 
 	// Generate shares for 5 participants
-	participantIDs := []int{1, 2, 3, 4, 5}
+	participantIDs := []int64{1, 2, 3, 4, 5}
 	shares := GenerateShares(poly, participantIDs)
 
 	if len(shares) != len(participantIDs) {
@@ -277,7 +277,7 @@ func testShareVerification(t *testing.T) {
 	}
 
 	// Generate shares
-	shares := GenerateShares(poly, []int{1, 2, 3, 4, 5})
+	shares := GenerateShares(poly, []int64{1, 2, 3, 4, 5})
 
 	// Verify all shares
 	for nodeID, share := range shares {
